changed GLSL bool value to java int
git-svn-id: http://fs.math.tu-berlin.de:8000/svn/jreality/trunk@2198 6a571d9c-c4e9-67b8-d2a0-cd0d85a3e679

+++ src-core/de/jreality/shader/GlslProgram.java
@@ -112,7 +112,7 @@ public class GlslProgram {
   public void setUniform(String name, boolean value) {
     checkWrite();
     checkParam(name, "bool", false, false);
-    assign(name, new float[]{value ? 1 : 0});
+    assign(name, new int[]{value ? 1 : 0});
   }
 
   public void setUniform(String name, float value) {
